+++ packages/yourloops/components/chat/chat-widget.tsx
@@ -42,7 +42,6 @@ import { Button, Tab, Tabs, TextField } from '@mui/material'
 import { useTranslation } from 'react-i18next'
 import { useTeam } from '../../lib/team'
 import { usePatientContext } from '../../lib/patient/patient.provider'
-import PatientUtils from '../../lib/patient/patient.util'
 import { type Patient } from '../../lib/patient/models/patient.model'
 import { UserRole } from '../../lib/auth/models/enums/user-role.enum'
 import { useSelectedTeamContext } from '../../lib/selected-team/selected-team.provider'
@@ -128,7 +127,7 @@ function ChatWidget(props: ChatWidgetProps): JSX.Element {
   const content = useRef<HTMLDivElement>(null)
   const inputRow = useRef<HTMLDivElement>(null)
   const { getUserName } = useUserName()
-  const teamId = user.isUserHcp() ? selectedTeam.id : PatientUtils.getRemoteMonitoringTeam(patient).teamId
+  const teamId = user.isUserHcp() ? selectedTeam.id : patient.teamIds[0]
 
   const handleChange = (_event: React.ChangeEvent, newValue: number): void => {
     setInputTab(newValue)
@@ -210,35 +209,41 @@ function ChatWidget(props: ChatWidgetProps): JSX.Element {
           </div>
         }
         <div id="chat-widget-footer" className={classes.chatWidgetFooter}>
-            {userRole === UserRole.Hcp &&
-              <Tabs
-                className={classes.chatWidgetTabs}
-                value={inputTab}
-                onChange={handleChange}
-                sx={{ marginBottom: theme.spacing(1) }}
-              >
-                <Tab
-                  className={classes.chatWidgetTab}
-                  label={t('chat-footer-reply')}
-                  aria-label={t('chat-footer-reply')}
-                  data-testid="chat-card-reply"
-                  onClick={() => { setPrivateMessage(false) }}
-                />
-                <Tab
-                  className={classes.chatWidgetTab}
-                  label={t('chat-footer-private')}
-                  aria-label={t('chat-footer-private')}
-                  data-testid="chat-card-private"
-                  onClick={() => { setPrivateMessage(true) }}
-                />
-              </Tabs>
-            }
+          {userRole === UserRole.Hcp &&
+            <Tabs
+              className={classes.chatWidgetTabs}
+              value={inputTab}
+              onChange={handleChange}
+              sx={{ marginBottom: theme.spacing(1) }}
+            >
+              <Tab
+                className={classes.chatWidgetTab}
+                label={t('chat-footer-reply')}
+                aria-label={t('chat-footer-reply')}
+                data-testid="chat-card-reply"
+                onClick={() => {
+                  setPrivateMessage(false)
+                }}
+              />
+              <Tab
+                className={classes.chatWidgetTab}
+                label={t('chat-footer-private')}
+                aria-label={t('chat-footer-private')}
+                data-testid="chat-card-private"
+                onClick={() => {
+                  setPrivateMessage(true)
+                }}
+              />
+            </Tabs>
+          }
           <div ref={inputRow} className={classes.chatWidgetInputRow}>
             <Button
               id="chat-widget-emoji-button"
               data-testid="chat-widget-emoji-button"
               className={classes.iconButton}
-              onClick={() => { setShowPicker(true) }}
+              onClick={() => {
+                setShowPicker(true)
+              }}
             >
               <SentimentSatisfiedOutlinedIcon />
             </Button>
@@ -249,7 +254,9 @@ function ChatWidget(props: ChatWidgetProps): JSX.Element {
               multiline
               maxRows={3}
               value={inputText}
-              onChange={event => { setInputText(event.target.value) }}
+              onChange={event => {
+                setInputText(event.target.value)
+              }}
               InputLabelProps={{ shrink: false }}
               data-testid="chat-card-input"
             />