+++ packages/yourloops/components/menus/team-scope-menu.tsx
@@ -52,6 +52,7 @@ import Divider from '@mui/material/Divider'
 import AddIcon from '@mui/icons-material/Add'
 import TeamEditDialog from '../../pages/hcp/team-edit-dialog'
 import { AppUserRoute } from '../../models/enums/routes.enum'
+import { usePatientContext } from '../../lib/patient/patient.provider'
 
 const classes = makeStyles()((theme: Theme) => ({
   sectionTitle: {
@@ -84,6 +85,7 @@ export const TeamScopeMenu: FunctionComponent = () => {
   const { getMedicalTeams, getPrivateTeam } = useTeam()
   const { selectTeam, selectedTeam } = useSelectedTeamContext()
   const { createTeam } = useTeam()
+  const { refresh } = usePatientContext()
   const alert = useAlert()
   const navigate = useNavigate()
   const [teamCreationDialogData, setTeamCreationDialogData] = React.useState<TeamEditModalContentProps | null>(null)
@@ -107,6 +109,7 @@ export const TeamScopeMenu: FunctionComponent = () => {
   const onSelectTeam = (teamId: string): void => {
     if (teamId !== selectedTeam.id) {
       selectTeam(teamId)
+      refresh(teamId)
 
       if (pathname !== AppUserRoute.Home) {
         navigate(AppUserRoute.Home)