+++ packages/yourloops/lib/patient/patient.api.ts
@@ -64,8 +64,8 @@ export default class PatientApi {
     }
   }
 
-  static async getPatientsForHcp(userId: string): Promise<Patient[]> {
-    const { data } = await HttpService.get<Patient[]>({ url: `/bff/v1/hcps/${userId}/patients` })
+  static async getScopedPatientsForHcp(userId: string, teamId: string): Promise<Patient[]> {
+    const { data } = await HttpService.get<Patient[]>({ url: `/bff/v1/hcps/${userId}/teams/${teamId}/patients` })
     return data
   }